docs: resolution for prettierJs issue
This was thought to be a prettier-js issue where lines were getting
realigned (more details in the commit diff). The reason this was
happening was because of the smartparens package. When hitting an open
parenthesis, a corresponding close parenthesis would be inserted. This
was also triggering a save which would run prettier-js because of the
onSave hook that was configured for it. This made prettier-js to save
the file mid-way and apply formatting with what it had.

Disabling smartparens helped fix the issue