refactor: moved individual configs to their respective folders
Learned this from DistroTube. Moving the configs to respective folders
makes them look cleaner in the git repo and more organized. Also,
renaming those files to README.org makes GitHub render them on the
"homepage" of the sub-directory. That way, it's one less of a step to
get to the readme.