nix: uninstall aerospace temporarily
There's an existing issue with aerospace and ghostty.
When a new tab is created in Ghostty, Aerospace adds an empty space in the workspace.
Changing tabs would move the Ghostty window from where it is to the empty space and back.

Basically, Ghostty implements tabs as native windows which is a MacOS behavior.
Aerospace uses MacOS APIs to treat windows as windows. So, Ghostty cannot really
do anything to fix it. Aerospace probably needs to find a workaround or a fix.
Until then, I won't use Aerospace because it gets extremely frustrating.